Fixed manual connection to PSU: Bad return device
The manual connection returned the previous device to the one chosen. The choice and connection were correct but not the return.
Also, it's checked the index selected is inside a port list length and converted to a number previously.

For example, if we want to select device 12, the function returned the device 11 and connected to device 12.